third_party/WebKit/LayoutTests/fast/html/imports/import-document-ready-state.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<body>
<script>
jsTestIsAsync = true;
successfullyParsed = true;

description('Tests that an imported document\'s readyState is "complete"');

function importLoaded() {
    shouldBe('i.import.readyState', '"complete"');
    finishJSTest();
}
</script>
<link id="i" rel="import" onload="importLoaded()" href="resources/hello.html" async>
