third_party/WebKit/LayoutTests/fast/html/imports/import-document-ready-state-expected.txt

Tests that an imported document's readyState is "complete"

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS i.import.readyState is "complete"
PASS successfullyParsed is true

TEST COMPLETE
